Feat/auth

+++ src/main/java/com/zeropick/backend/comparison/dto/ProductCompareResponse.java
@@ -17,7 +17,7 @@ public class ProductCompareResponse {
     public static class ComparisonItem {
         private Long productId;
         private String productName;
-        private Integer grade;
+        private Short grade;
         private Boolean warningAdditive;
         private Nutrition nutrition;
         private KeyIngredients keyIngredients;
